Moved code into the dstu2 directory Updated all the paths Updated all the configs
