+++ src/config.js
@@ -5,7 +5,6 @@
 module.exports.files = {
 	swaggerHelperDefinitions: 'src/server/*/definitions/*.json',
 	swaggerRootDefinitions: 'src/server/*/*.swagger.json',
-	conformanceStatements: 'src/server/*/conformance.js',
 	routes: 'src/server/**/*.routes.js'
 };
