+++ src/server/dstu2/metadata/routes/metadata.routes.js
@@ -1,5 +1,5 @@
 const cors = require('cors');
-const { sanitizeMiddleware } = require('../../utils/sanitize.utils');
+const { sanitizeMiddleware } = require('../../../utils/sanitize.utils');
 const { routes } = require('../metadata.config');
 
 /**
