+++ src/server/dstu2/observation/routes/observation.routes.js
@@ -1,7 +1,8 @@
 const cors = require('cors');
-const { sanitizeMiddleware } = require('../../utils/sanitize.utils');
-const { validate } = require('../../utils/auth');
-const { routes } = require('../patient.config');
+const { sanitizeMiddleware } = require('../../../utils/sanitize.utils');
+const { routes } = require('../observation.config');
+const { validate } = require('../../../utils/auth');
+
 
 /**
  * @name exports
@@ -9,25 +10,24 @@ const { routes } = require('../patient.config');
  */
 module.exports = (app, config, logger) => {
 	let { profiles, server } = config;
+	let observationProfile = profiles.dstu2 && profiles.dstu2.observation;
 
-	// Only add routes if we have a patient profile
-	// the endpoint can't function without the config
-	if (profiles.patient) {
-		let baseOptions = Object.assign({}, server.corsOptions, profiles.patient.corsOptions);
+	if (observationProfile) {
+		let baseOptions = Object.assign({}, server.corsOptions, observationProfile.corsOptions);
 
 		routes.forEach((route) => {
 			let corsOptions = Object.assign({}, baseOptions, route.corsOptions);
 			// Enable options
 			app.options(route.path, cors(corsOptions));
 			// Enable route
+
 			app[route.type](
 				route.path,
 				cors(corsOptions),
 				sanitizeMiddleware(route.args),
 				validate(route.scopes, logger, config),
-				route.controller(profiles.patient, logger)
+				route.controller(observationProfile, logger)
 			);
 		});
 	}
-
 };
