+++ src/server/dstu2/patient/controllers/patient.controller.js
@@ -1,4 +1,4 @@
-const { ServerError } = require('../../utils/error.utils');
+const { ServerError } = require('../../../utils/error.utils');
 
 module.exports.getPatient = (profile, logger) => {
 	let { serviceModule: service } = profile;
