+++ src/server/dstu2/patient/routes/patient.routes.js
@@ -1,8 +1,7 @@
 const cors = require('cors');
-const { sanitizeMiddleware } = require('../../utils/sanitize.utils');
-const { routes } = require('../observation.config');
-const { validate } = require('../../utils/auth');
-
+const { sanitizeMiddleware } = require('../../../utils/sanitize.utils');
+const { validate } = require('../../../utils/auth');
+const { routes } = require('../patient.config');
 
 /**
  * @name exports
@@ -10,24 +9,26 @@ const { validate } = require('../../utils/auth');
  */
 module.exports = (app, config, logger) => {
 	let { profiles, server } = config;
+	let patientProfile = profiles.dstu2 && profiles.dstu2.patient;
+
 	// Only add routes if we have a patient profile
 	// the endpoint can't function without the config
-	if (profiles.observation) {
-		let baseOptions = Object.assign({}, server.corsOptions, profiles.observation.corsOptions);
+	if (patientProfile) {
+		let baseOptions = Object.assign({}, server.corsOptions, patientProfile.corsOptions);
 
 		routes.forEach((route) => {
 			let corsOptions = Object.assign({}, baseOptions, route.corsOptions);
 			// Enable options
 			app.options(route.path, cors(corsOptions));
 			// Enable route
-
 			app[route.type](
 				route.path,
 				cors(corsOptions),
 				sanitizeMiddleware(route.args),
 				validate(route.scopes, logger, config),
-				route.controller(profiles.observation, logger)
+				route.controller(patientProfile, logger)
 			);
 		});
 	}
+
 };
