+++ src/server/dstu2/resources/index.js
@@ -1,0 +1,17 @@
+const Observation = require('../../resources/Observation');
+const {
+	Patient,
+	Communication,
+	PatientContact
+} = require('../../resources/Patient');
+
+/**
+ * @name exports
+ * @description Export all the DSTU2 resource types
+ */
+module.exports = {
+	Observation,
+	Patient,
+	Communication,
+	PatientContact
+};
